appPython/requirements.txt
click==6.7
Flask==0.12.2
gunicorn==19.7.1
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
PyJWT==1.7.1
virtualenv==15.1.0
Werkzeug==0.12.2
